refactor: extract input overlay helpers
